requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Parameter-efficient fine-tuning
peft>=0.6.0
bitsandbytes>=0.41.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Evaluation and metrics
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0

# Data collection
requests>=2.31.0
beautifulsoup4>=4.12.0
PyGithub>=2.1.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Logging and monitoring
wandb>=0.16.0
tensorboard>=2.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
